pyproject.toml
[project]
name = "move"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=0.20.0",
    "bitsandbytes>=0.39.0",
    "datasets>=2.12.0",
    "huggingface-hub>=0.15.0",
    "ipykernel>=6.23.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "peft>=0.4.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "seaborn>=0.12.0",
    "tensorboard>=2.13.0",
    "tokenizers>=0.13.0",
    "torch-geometric>=2.3.0",
    "tqdm>=4.65.0",
    "transformers>=4.30.0",
    "wandb>=0.15.0",
]
